+++ android/src/main/java/com/lodev09/truesheet/TrueSheetPackage.kt
@@ -40,6 +40,7 @@ class TrueSheetPackage : BaseReactPackage() {
       TrueSheetContentViewManager(),
       TrueSheetHeaderViewManager(),
       TrueSheetFooterViewManager(),
-      TrueSheetPeekViewManager()
+      TrueSheetPeekViewManager(),
+      TrueSheetOverlayViewManager()
     )
 }