+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './TrueSheet';
 export * from './TrueSheet.types';
 export { TrueSheetPeek } from './TrueSheetPeek';
+export { TrueSheetOverlay } from './TrueSheetOverlay';
 export { TrueSheetProvider, useTrueSheet } from './TrueSheetProvider';